requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==2.1.1
beautifulsoup4==4.12.2
requests==2.31.0
nltk==3.8.1
scikit-learn==1.7.1
scipy==1.16.1
feedparser==6.0.11
PyYAML==6.0.1
